+++ src/pages/Save.tsx
@@ -2,11 +2,11 @@ import { isMobile } from 'react-device-detect';
 import SaveHeader from '@/components/layout/header/SaveHeader';
 import { useNavigate } from 'react-router-dom';
 import { tv } from 'tailwind-variants';
-import CategoryTagSelector from '@/pages/save/CategoryTagSelector';
-import Memo from '@/pages/save/Memo';
-import Alarm from '@/pages/save/Alarm';
+import CategoryTagSelector from '@/components/ui/cardlink/CategoryTagSelector';
+import Memo from '@/components/ui/cardlink/Memo';
+import Alarm from '@/components/ui/cardlink/Alarm';
 import Button from '@/components/common/Button';
-import LinkField from '@/pages/save/LinkField';
+import LinkField from '@/components/ui/cardlink/LinkField';
 import { isSaveButtonDisabledAtom } from '@/atoms';
 import { useAtomValue } from 'jotai';
 import { useScrollLock } from '@/components/hooks/ScrollLock';
