+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import Home from './Home';
 import Login from './Login';
-import Save from './save/Save';
+import Save from './Save';
 import Search from './Search';
 import Example from './Example';
 import SearchResult from './SearchResult';
